+++ content-featured-post.php
@@ -14,7 +14,10 @@ if ( ! defined( 'ABSPATH' ) ) {
 <article id="post-<?php the_ID(); ?>" <?php post_class(); ?>>
 	<?php if ( has_post_thumbnail() ) : ?>
 		<a class="post-thumbnail" href="<?php echo esc_url( get_permalink() ); ?>">
-			<?php the_post_thumbnail( 'canard-featured-content-thumbnail', array( 'loading' => 'lazy' ) ); ?>
+			<?php the_post_thumbnail( 'canard-featured-content-thumbnail', array(
+			'loading' => 'lazy',
+			'sizes'   => '(max-width: 1300px) 100vw, 1300px',
+		) ); ?>
 		</a>
 	<?php endif; ?>